+++ src/Application/Authorization/EventCapabilityGate.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace EventFlow\Application\Authorization;
+
+use EventFlow\Application\Persistence\EventScope;
+
+interface EventCapabilityGate
+{
+    public function allows(EventScope $scope, Capability $capability): bool;
+}