+++ src/Application/Authorization/PermissiveEventCapabilityGate.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace EventFlow\Application\Authorization;
+
+use EventFlow\Application\Persistence\EventScope;
+
+final readonly class PermissiveEventCapabilityGate implements EventCapabilityGate
+{
+    public function allows(EventScope $scope, Capability $capability): bool
+    {
+        return true;
+    }
+}